Replace uint with uint256 in stake function parameter

+++ StakingContract/Rebase.sol
@@ -1980,7 +1980,7 @@ contract Rebase is ReentrancyGuard {
 
     receive() external payable { }
 
-    function stake(address token, uint quantity, address app) external nonReentrant {
+    function stake(address token, uint256 quantity, address app) external nonReentrant {
         require(ERC20(token).transferFrom(msg.sender, address(this), quantity), "Unable to transfer token");
         _getReToken(token).mint(msg.sender, quantity);
         _stake(app, token, quantity);
